feat: Add toJson method to RestaurantDTO
The code changes add a new method called `toJson` to the `RestaurantDTO` class in the `restaurant_dto.dart` file. This method converts the `RestaurantDTO` object into a JSON representation by mapping its properties to a `Map<String, dynamic>`. This addition enhances the functionality of the application by providing a convenient way to serialize the `RestaurantDTO` object for network communication or storage purposes.