feat: Update logout navigation to use pushReplacement
This commit updates the logout navigation in the `logout.dart` file to use `Navigator.pushReplacement` instead of `Navigator.push`. This change ensures that the login page is displayed as the new route after logging out, replacing the previous route in the navigation stack. This enhancement improves the user experience by preventing the user from navigating back to the previous page after logging out.