feat: Add controller to TextFieldWLabel widget
This commit adds a new `controller` property to the `TextFieldWLabel` widget in the `modal_form_text_field.dart` file. The `controller` allows for better control and manipulation of the text entered in the text field. This enhancement improves the functionality of the widget by providing a convenient way to access and modify the text value.