feat: Add Equatable to admin state and user models
This commit adds the `Equatable` mixin to the `AdminState` and `CustomerModel` classes in order to improve the equality comparison and hashing of objects. By implementing `Equatable`, the `props` list is defined, which includes the necessary properties for comparison. This enhancement ensures more accurate equality checks and improves the overall reliability of the code.